backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.5,<6.0
django-cors-headers>=4.9,<5.0
python-dotenv>=1.2,<2.0
psycopg2-binary>=2.9,<3.0
gunicorn>=23.0,<24.0
